ai/requirements.txt
pandas>=2.2.0
scikit-learn>=1.5.0
joblib>=1.4.0
mysql-connector-python>=9.0.0
numpy>=1.26.0
